WIP: re organized directory structure, revising docker-compose file so that a particular service, and its dependencies, can be simply started up
