+++ docker-compose.yml
@@ -2,9 +2,15 @@
 # It will 'tie' together the DB, Preprocessor, Frontend, Skill Oracle Dockerfiles
 # and present an easy to use (e.g., `docker-compse up`) interface for standing
 # the entire system up on a personal system
+#
 version: '3'
 services:
   skilloracle:
     image: skilloracle:latest
-    ports:
-        - "4000:4000"
+    build:
+      context: ./
+      dockerfile: ./skilloracle/Dockerfile
+    depends-on:
+      - redis
+  redis:
+    image: redis
