+++ skilloracle/api/api.py
@@ -1,4 +1,4 @@
-""" Skill Labeller Preprocessor API """
+""" Skill Labeller Skill Oracle API """
 
 import falcon
 from ..endpoint.skilloracle import SkillOracleEndpoint
